move the builds bit into a module
